requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters

sentence-transformers
faiss-cpu
torch
torchvision

pandas
PyPDF2
python-dotenv
streamlit

fastapi
uvicorn[standard]

python-multipart

langfuse